requirements.txt
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
biomart==0.9.2
biopython==1.85
biotite==0.41.2
biotraj==1.2.2
boto3==1.40.2
botocore==1.40.2
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
cloudpathlib==0.21.1
Cython==3.1.2
decorator==5.2.1
einops==0.8.1
esm==3.1.2
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
hdbscan==0.8.38.post1
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
lmdb==1.7.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
msgpack-numpy==0.4.8
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tape-proteins==0.5
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.7.1
torchtext==0.18.0
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.2
triton==3.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.5.0
wcwidth==0.2.13
zstd==1.5.7.2
